+++ packages/flutter/lib/src/native/cocoa/sentry_native_cocoa.dart
@@ -1,6 +1,8 @@
 import 'dart:async';
 import 'dart:ffi';
 
+import 'package:ffi/ffi.dart';
+import 'package:flutter/foundation.dart';
 import 'package:meta/meta.dart';
 
 import '../../../sentry_flutter.dart';
@@ -50,6 +52,34 @@ class SentryNativeCocoa extends SentryNativeChannel {
     }
 
     return super.init(hub);
+  }
+
+  @override
+  FutureOr<void> captureEnvelope(
+      Uint8List envelopeData, bool containsUnhandledException) {
+    // Use a safe copy-based conversion to avoid crashes due to memory issues observed
+    // when relying on `dataWithBytesNoCopy:length:freeWhenDone:`.
+    final length = envelopeData.length;
+    final ptr = malloc<Uint8>(length);
+    ptr.asTypedList(length).setAll(0, envelopeData);
+    try {
+      final nsData =
+          cocoa.NSData.dataWithBytes_length_(_lib, ptr.cast<Void>(), length);
+
+      final envelope =
+          cocoa.PrivateSentrySDKOnly.envelopeWithData_(_lib, nsData);
+
+      cocoa.PrivateSentrySDKOnly.captureEnvelope_(_lib, envelope);
+    } catch (exception, stackTrace) {
+      options.log(SentryLevel.error, 'Failed to capture envelope',
+          exception: exception, stackTrace: stackTrace);
+
+      if (options.automatedTestMode) {
+        rethrow;
+      }
+    } finally {
+      malloc.free(ptr);
+    }
   }
 
   @override